Fix for Grid.GetCellId
Row and col parameters for Grid.GetCellId are passed by value and not by reference, so the existing function only partially worked. This change properly implements functions to get the cellid from row and col values or from coordinates and a function to get row and col values from coordinates.