Waypoint flowfield with visualization
First experimenting with something like flowfields (grid-based here). This one starts at the next waypoint and calculates a number for the shortest distance to the waypont for all cells. The idea is that units in formation will only move to cells with equal or lower numbers, prefering those with lower numbers.